ex3.py

+++ ex3.py
@@ -1,3 +1,4 @@
-ng = float(input("número multiplicado:"))
+ng = int(input("número multiplicado:"))
 for nt in range(10):
+
  print(ng, "X", 1+nt,"=", ((1+nt)*ng))
